+++ js/components/SkinsContextMenu.js
@@ -1,12 +1,13 @@
 import React from "react";
 import { connect } from "react-redux";
-import { setSkinFromUrl, openSkinFileDialog } from "../actionCreators";
+import * as Actions from "../actionCreators";
 import { Hr, Node, Parent } from "./ContextMenu";
 
 const SkinContextMenu = props => (
   <Parent label="Skins">
     <Node onClick={props.openSkinFileDialog} label="Load Skin..." />
-    {!!props.availableSkins.length && <Hr />}
+    <Hr />
+    <Node onClick={props.loadDefaultSkin} label={"<Base Skin>"} />
     {props.availableSkins.map(skin => (
       <Node
         key={skin.url}
@@ -22,8 +23,9 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = {
-  openSkinFileDialog,
-  setSkin: setSkinFromUrl
+  loadDefaultSkin: Actions.loadDefaultSkin,
+  openSkinFileDialog: Actions.openSkinFileDialog,
+  setSkin: Actions.setSkinFromUrl
 };
 
 export default connect(